[refactor] 구간 추가 로직 리팩토링
