+++ src/main/java/nextstep/line/Line.java
@@ -5,10 +5,9 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import nextstep.exception.InvalidInputException;
-import nextstep.section.Section;
-import nextstep.section.Sections;
+import nextstep.line.section.Section;
+import nextstep.line.section.Sections;
 import nextstep.station.Station;
-import org.jgrapht.alg.util.Pair;
 
 import javax.persistence.*;
