+++ src/main/java/nextstep/line/section/SectionAddRequest.java
@@ -1,4 +1,4 @@
-package nextstep.section;
+package nextstep.line.section;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
